requirements.txt
flask
requests
python-dotenv
pytesseract
Pillow